chore: move code to separate files
